test(otel): fix shared fixture plugin collection
